✨ Feat : 엑세스 토큰 사용 - 유저 정보 조회 API 구현

+++ src/main/java/com/example/gguro/repository/UserRepository.java
@@ -1,14 +1,13 @@
 package com.example.gguro.repository;
 
 import com.example.gguro.domain.User;
-import jakarta.validation.constraints.NotBlank;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
     boolean existsByUsername(String username);
     Optional<User> findByUsername(String username);
 }
